Refactoring while I look for this bug
